tagging/requirements.txt
langchain
langchain_community
langchain_core
langchain_openai
python-dotenv
pandas
bs4